Add script to update openmediavault
Because in later OMV releases the openmediavault package repositories are no
longer pinned with a high value (995) by default setting APT::Default-Release
pins the Debian repository higher. So some dependencies for omv5 might not be
installed and omv5 will not be upgraded. With this step we do a distinct
upgrade attempt.

Fixes #7